reorganize mixins
