--- src/main/java/io/github/marinersfan824/racemod/ILevelProperties.java
@@ -1,5 +1,0 @@
-package io.github.marinersfan824.racemod;
-
-public interface ILevelProperties {
-    RNGStreamGenerator getRngStreamGenerator();
-}
