+++ src/main/java/io/github/marinersfan824/racemod/mixin/worldgen/CaveMixin.java
@@ -1,4 +1,4 @@
-package io.github.marinersfan824.racemod.mixin;
+package io.github.marinersfan824.racemod.mixin.worldgen;
 
 import net.minecraft.client.font.TextRenderer;
 import net.minecraft.client.gui.hud.InGameHud;
