+++ src/main/java/io/github/marinersfan824/racemod/mixin/worldgen/RavineMixin.java
@@ -1,4 +1,4 @@
-package io.github.marinersfan824.racemod.mixin;
+package io.github.marinersfan824.racemod.mixin.worldgen;
 
 import net.minecraft.block.Block;
 import net.minecraft.world.gen.carver.Carver;
