+++ src/main/java/io/github/marinersfan824/racemod/mixin/worldgen/SpiralStaircaseMixin.java
@@ -1,10 +1,9 @@
-package io.github.marinersfan824.racemod.mixin;
+package io.github.marinersfan824.racemod.mixin.worldgen;
 
-import io.github.marinersfan824.racemod.ISpiralStaircase;
+import io.github.marinersfan824.racemod.mixinterface.ISpiralStaircase;
 import net.minecraft.class_26;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.util.math.BlockBox;
-import net.minecraft.util.math.Vec3i;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
