+++ src/main/java/io/github/marinersfan824/racemod/mixin/worldgen/StartPieceMixin.java
@@ -1,6 +1,6 @@
-package io.github.marinersfan824.racemod.mixin;
+package io.github.marinersfan824.racemod.mixin.worldgen;
 
-import io.github.marinersfan824.racemod.ISpiralStaircase;
+import io.github.marinersfan824.racemod.mixinterface.ISpiralStaircase;
 import net.minecraft.class_27;
 import net.minecraft.util.math.BlockBox;
 import net.minecraft.util.math.Vec3i;
