+++ src/main/java/io/github/marinersfan824/racemod/mixinterface/ILevelProperties.java
@@ -1,0 +1,7 @@
+package io.github.marinersfan824.racemod.mixinterface;
+
+import io.github.marinersfan824.racemod.RNGStreamGenerator;
+
+public interface ILevelProperties {
+    RNGStreamGenerator getRngStreamGenerator();
+}
