+++ src/main/java/io/github/marinersfan824/racemod/mixinterface/ISpiralStaircase.java
@@ -1,4 +1,4 @@
-package io.github.marinersfan824.racemod;
+package io.github.marinersfan824.racemod.mixinterface;
 
 import net.minecraft.util.math.BlockBox;
