+++ src/main/resources/racemod.mixins.json
@@ -5,23 +5,23 @@
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "BedrockDropMixin",
-    "EndPortalRoomMixin",
-    "EyeOfEnderMixin",
-    "LakesMixin",
-    "LevelPropertiesMixin",
-    "TitleScreenMixin",
-    "MobDropsMixin",
-    "RavineMixin",
-    "SpiralStaircaseMixin",
-    "StartPieceMixin",
-    "TitleScreenMixin",
-    "CaveMixin",
-    "TempleMobSpawnMixin",
-    "ChunkSectionAccessor",
-    "FlintMixin",
-    "BiomeMixin",
-    "NetherFortressStructureMixin",
-    "SurfaceChunkGeneratorMixin"
+    "worldgen.EndPortalRoomMixin",
+    "rng.EyeOfEnderMixin",
+    "worldgen.LakesMixin",
+    "rng.LevelPropertiesMixin",
+    "misc.TitleScreenMixin",
+    "rng.MobDropsMixin",
+    "worldgen.RavineMixin",
+    "worldgen.SpiralStaircaseMixin",
+    "worldgen.StartPieceMixin",
+    "misc.TitleScreenMixin",
+    "worldgen.CaveMixin",
+    "rng.TempleMobSpawnMixin",
+    "access.ChunkSectionAccessor",
+    "rng.FlintMixin",
+    "worldgen.BiomeMixin",
+    "worldgen.NetherFortressStructureMixin",
+    "worldgen.SurfaceChunkGeneratorMixin"
   ],
   "client": [
   ],
